backend/requirements.txt
Flask>=2.0
Flask-CORS>=3.0
Flask-SQLAlchemy>=2.5
gunicorn>=20.0  # Needed for production servers like Heroku/Render
python-dotenv>=1.0  # If you use .env for secret keys
google-generativeai>=0.2.2  # Replace with exact version you're using